Fix spectral folder rename crash and arrow keys in docker

+++ salmon/tagger/foldername.py
@@ -85,13 +85,21 @@ def rename_folder(path, metadata, auto_rename, check=True):
         tmp_old_specs_path = os.path.join(cfg.directory.tmp_dir, f"spectrals_{old_base}")
         tmp_new_specs_path = os.path.join(cfg.directory.tmp_dir, f"spectrals_{new_base}")
 
-        if os.path.exists(tmp_old_specs_path):
-            if os.path.exists(tmp_new_specs_path) and not os.path.samefile(tmp_old_specs_path, tmp_new_specs_path):
-                shutil.rmtree(tmp_new_specs_path)
-
+        if (
+            os.path.exists(tmp_old_specs_path)
+            and os.path.exists(tmp_new_specs_path)
+            and os.path.samefile(tmp_old_specs_path, tmp_new_specs_path)
+        ):
+            click.secho(f"Skipping copy, same location already for '{tmp_new_specs_path}'", fg="yellow")
+        else:
             if use_hardlinks:
-                shutil.copytree(tmp_old_specs_path, tmp_new_specs_path, copy_function=os.link, dirs_exist_ok=True)
-                click.secho(f"Hardlinked temporary spectrals folder to '{tmp_new_specs_path}'.", fg="yellow")
+                try:
+                    shutil.copytree(tmp_old_specs_path, tmp_new_specs_path, copy_function=os.link, dirs_exist_ok=True)
+                    click.secho(f"Hardlinked temporary spectrals folder to '{tmp_new_specs_path}'.", fg="yellow")
+                except shutil.Error as _:
+                    click.secho("Hardlinking didn't work, falling back to non-hardlink copy...", fg="red")
+                    shutil.copytree(tmp_old_specs_path, tmp_new_specs_path, dirs_exist_ok=True)
+                    click.secho(f"Copied temporary spectrals folder to '{tmp_new_specs_path}'.", fg="yellow")
             else:
                 shutil.copytree(tmp_old_specs_path, tmp_new_specs_path, dirs_exist_ok=True)
                 click.secho(f"Copied temporary spectrals folder to '{tmp_new_specs_path}'.", fg="yellow")
